requirements.txt
Flask==3.1.2
gunicorn==25.1.0
flask-sqlalchemy==3.1.1
pymysql==1.1.2
python-dotenv==1.0.1
cryptography==46.0.5
Flask-Migrate==4.1.0
argon2-cffi==25.1.0